refactor: :truck: rename MainLayout -> BlogLayout
