+++ src/components/common/layouts/blog-layout/index.style.ts
